+++ client/src/components/timeline-create-entry-components/note-entry-component/index.jsx
@@ -1,15 +1,10 @@
-import { useState } from "react";
-import { Card, Box, Avatar, CardContent, TextField, Button, Typography, Divider, Select, MenuItem } from "@mui/material";
-import PhoneIphoneIcon from '@mui/icons-material/PhoneIphone';
+import { useState } from "react"
+import { Card, Box, Avatar, CardContent, TextField, Button, Typography, Divider } from "@mui/material";
+import CreateIcon from '@mui/icons-material/Create';
 // import CloseIcon from '@mui/icons-material/Close';
 
-function EmployerEntryComponent() {
-    const [selectedValue, setSelectedValue] = useState('Employer');
+function NoteEntryComponent() {
     const [notes, setNotes] = useState("")
-
-    const handleValueChange = (e) => {
-      setSelectedValue(e.target.value);
-    };
 
     const handleNoteChange = (e) => {
         setNotes(e.target.value)
@@ -21,34 +16,24 @@ function EmployerEntryComponent() {
                 <Typography variant="h5">Activity Timeline</Typography>
                 <Box sx={{ display: "flex", marginTop: "12px" }}>
                     <Box>
-                        <Avatar sx={{ bgcolor: "rgba(229, 53, 104, 1)", width: "30px", height: "30px" }}>
-                            <PhoneIphoneIcon sx={{ width: "18px", height: "18px" }}/>
+                        <Avatar sx={{ bgcolor: "rgba(255, 180, 0, 1)", width: "30px", height: "30px" }}>
+                            <CreateIcon sx={{ width: "18px", height: "18px" }}/>
                         </Avatar>
                     </Box>
                     <Box sx={{ display: "flex", justifyContent: "center", alignItems: "center", marginLeft: "16px" }}>
-                <Typography>Contacted Employer</Typography>
+                <Typography>Write a Note</Typography>
                     </Box>
                 </Box>
             </CardContent>
             <Divider />
             <CardContent>
-            <Typography sx={{ color: "rgba(117, 117, 117, 1)" }}>Select the Employer Contact</Typography>
-                <Select
-                 fullWidth
-                 value={selectedValue}
-                 onChange={handleValueChange}
-                 label="Choose an option"
-                 sx={{ borderRadius: '10px' }} 
-                >
-                <MenuItem value="Employer">Employer</MenuItem>
-                </Select>
                 <TextField
                     fullWidth
-                    label="Add additional notes"
+                    label="Write your note"
                     margin="normal"
                     multiline
                     variant="outlined"
-                    rows={20}
+                    rows={23}
                     value={notes}
                     onChange={handleNoteChange}
                 />
@@ -61,4 +46,4 @@ function EmployerEntryComponent() {
     )
 }
 
-export default EmployerEntryComponent;
+export default NoteEntryComponent;
